Corrected Error with modify after directly adding item
Whenever the user would add an item. Should they have tried to modify the item after creating it, it would encounter an error stating the item was not in the inventory. This was an error of using add item function which stored item num as a int and not a string.